# [TIMOB-26331] Improve build/launch time via Titanium CLI