--- a/cli/hooks/ws-run.js
+++ b/cli/hooks/ws-run.js
@@ -157,7 +157,7 @@ exports.init = function (logger, config, cli) {
 					appId = tiapp.windows.id || tiapp.id,
 					sanitizedName = builder.sanitizeProjectName(tiapp.name),
 					// name of the directory holding appx and dependencies subfolder
-					dirName = sanitizedName + '_' + builder.buildVersion + ((builder.buildConfiguration == 'Debug') ? '_Debug_Test' : '_Test'),
+					dirName = sanitizedName + '_' + builder.buildVersion + (builder.useAppxBundle ? '' : ('_' + builder.cmakeArch)) + ((builder.buildConfiguration == 'Debug') ? '_Debug_Test' : '_Test'),
 					projectDir = path.resolve(builder.cmakeTargetDir, 'AppPackages', sanitizedName, dirName),
 					// Options for installing and launching app
 					opts = appc.util.mix({