[TIMOB-26331](https://jira.appcelerator.org/browse/TIMOB-26331)

Attempt to improve build/launch time performance via CLI.

## Things that work

- Reduced re-build time on device by stopping unnecessary clean build 
- Reduced development build (local/device) time by stopping unnecessary bundle package
- Reduced build time by stopping unnecessary `cmake` invocation from `msbuild`
- Reduced build time by stopping unnecessary compilation for C++ files

Overall this reduces 15-30 seconds build/launch time (40~% improvement).

## Things that do not work

- Moving to native [MSBuild API](https://docs.microsoft.com/en-us/visualstudio/msbuild/msbuild-api?view=vs-2017) instead of invoking `msbuild` command does not quite work to reduce build time.
- Moving to native [package deployment API](https://docs.microsoft.com/en-us/windows/desktop/appxpkg/package-deployment-api) instead of invoking `WinAppDeployCmd` command does not quite work to reduce deployment time.

## Another solution regarding development "productivity"

I would suggest leveraging `liveview` rather than recompiling projects in order to improve development productivity. I even expect it works faster than Visual Studio to preview your changes!

## CLEAN BUILD

### BEFORE

```
[INFO]  Project built successfully in 1m 9s 450ms
[INFO]  Finished launching the application in 1m 27s 387ms
```

### AFTER

```
[INFO]  Project built successfully in 55s 204ms
[INFO]  Finished launching the application in 1m 10s 142ms
```

## REBUILD

### BEFORE

```
[INFO]  Project built successfully in 47s 360ms
[INFO]  Finished launching the application in 1m 8s 970ms
```

### AFTER

```
[INFO]  Project built successfully in 27s 715ms
[INFO]  Finished launching the application in 40s 950ms
```